backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
mediapipe>=0.10.0
exifread>=3.0.0
cryptography>=41.0.0
redis>=5.0.0
gunicorn>=21.2.0
